+++ app/templates/gulp/custom/styles.js
@@ -1,6 +1,7 @@
 var gulp = require('gulp'),
   <%= styles %> = require('gulp-<%= styles %>'),
-  config = require('../config/build.conf');
+  config = require('../config/build.conf'),
+  logger = console;
 
 /**
  * Tasks to preprocess <%= styles %> files.
@@ -9,6 +10,10 @@ gulp.task('<%= styles %>', function() {
   return gulp.src([
     config.source.app + '/styles/styles<% if (styles === 'less') { %>.less'<% } else if (styles === 'sass') { %>.scss'<% } else if (styles === 'stylus') { %>.styl'<% } %>
   ])
-  .pipe(<%= styles %>())
+  .pipe(<%= styles %>().on('error', function(error) {
+    logger.error(error.message);
+
+    this.emit('end');
+  }))
   .pipe(gulp.dest(config.source.tmp + '/styles'));
 });