+++ app/templates/gulp/tasks/concat.js
@@ -7,7 +7,7 @@ var gulp = require('gulp'),
  */
 gulp.task('concat:js', function() {
   return gulp.src([
-    '!' + config.pkg.name + '.js',
+    '!' + config.source.tmp + '/' + config.pkg.name + '.js',
     config.source.tmp + '/**/*.js'
   ])
   .pipe(concat(config.pkg.name + '.js'))
@@ -19,7 +19,7 @@ gulp.task('concat:js', function() {
  */
 gulp.task('concat:css', function() {
   return gulp.src([
-    '!' + config.pkg.name + '.css',
+    '!' + config.source.tmp + '/' + config.pkg.name + '.css',
     config.source.tmp + '/**/*.css'
   ])
   .pipe(concat(config.pkg.name + '.css'))